% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndreevShkolnikovBegun.R
\name{LTabr}
\alias{LTabr}
\title{an abridged lifetable based on M(x)}
\usage{
LTabr(Mx, Age = c(0, 1, cumsum(rep(5, length(Mx) - 2))), radix = 1e+05)
}
\arguments{
\item{Mx}{numeric vector of abridged mortality rates.}

\item{Age}{integer, abrdged age lower bounds.}

\item{radix}{numeric. Can be anything positive.}
}
\value{
numeric life expectancy at birth
}
\description{
Implements the abridged lifetable formulas given in the supplementary material to Andreev et al (2012). An entire lifetable is calculated, but only lfie expectancy at birth is returned.
}
\details{
Chiang's a(x) is assumed in the following way: \eqn{a(0) = 0.07 + 1.7 * M(0)}, \eqn{a(1) = 1.6}, \eqn{a(\omega) = \frac{1}{M(\omega)}}, and all others are assumed at mid interval. The last age is assumed open. Everything else is pretty standard.
}
\references{
E. Andreev, V. Shkolnikov, and A. Begun (2002) Algorithm for decomposition of differences between aggregate demographic measures and its application to life expectancies, healthy life expectancies, parity-progression ratios and total fertility rates. Demographic Research v7, n14

E. Andreev, V. Shkolnikov, and A. Begun (2012) An Excel spreadsheet for the decomposition of a difference between two values of an aggregate demographic measure by stepwise replacement running from young to old ages. MPIDR Technical Report TR-2012-002
}
